requirements.txt
art==4.7
colorama==0.4.3
filetype==1.0.7
future==0.18.2
numpy==1.19.0
pandas==1.0.5
pefile==2019.4.18
python-dateutil==2.8.1
pytz==2020.1
rarfile==3.1
six==1.15.0
tqdm==4.47.0
yara-python==4.0.2
